requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
